refactor: 매서드를 별개의 파일로 분리
