+++ src/index.ts
@@ -1,3 +1,17 @@
+import _fetch from "./fetch";
+import {
+  isNull as _isNull,
+  isNil as _isNil,
+  isNumber as _isNumber,
+  isFunction as _isFunction,
+} from "./checkType";
+import _shuffle from "./shuffle";
+import _pick from "./pick";
+import _omit from "./omit";
+import _memoize from "./memoize";
+import _debounce from "./debounce";
+import _throttle from "./throttle";
+import _clickOutside from "./clickOutside";
 interface CustomElementMethod {
   addEvent<T extends keyof GlobalEventHandlersEventMap>(
     eventType: T,
@@ -85,256 +99,18 @@ function _(selector: string) {
 }
 
 module _ {
-  /**
-   * XMLHttpRequest를 활용해 비동기적으로 서버와 통신합니다.
-   *
-   * @param url 요청을 받을 서버의 주소
-   * @param method 요청 매서드
-   * @param payload 요청 body에 추가할 정보
-   * @returns 서버로부터 받은 응답
-   */
-  export function fetch(
-    url: string,
-    method: string,
-    payload?: unknown
-  ): Promise<any> {
-    const request = new XMLHttpRequest();
-    request.open(method, url);
-    request.responseType = "json";
-    const requestBody = JSON.stringify(payload);
-
-    let resolve: unknown, reject: unknown;
-
-    request.onload = function (this: XMLHttpRequest) {
-      if (typeof resolve !== "function")
-        throw new Error("잘못 설정된  resolver");
-      return resolve(this.response);
-    };
-
-    request.onerror = function (this: XMLHttpRequest) {
-      if (typeof reject !== "function")
-        throw new Error("잘못 설정된  rejecter");
-      return reject(this.response);
-    };
-
-    request.send(requestBody);
-
-    return new Promise((res, rej) => {
-      resolve = res;
-      reject = rej;
-    });
-  }
-
-  /**
-   * null인지 여부를 반환합니다.
-   *
-   * @param input 검증 대상 값
-   * @returns input이 null인지 여부
-   */
-  export function isNull<T>(input: T): T extends null ? true : false;
-  export function isNull(input: unknown): input is null {
-    return input === null;
-  }
-
-  /**
-   *  null 혹은 undefined인지 여부를 반환합니다.
-   *
-   * @param input 검증 대상 값
-   * @returns input이 null 혹은 undefined인지 여부
-   */
-  export function isNil<T>(input: T): T extends null | undefined ? true : false;
-  export function isNil(input: unknown): input is null | undefined {
-    return input === null || input === undefined;
-  }
-
-  /**
-   * number인지 여부를 반환합니다.
-   *
-   * @param input 검증 대상 값
-   * @returns input이 number인지 여부
-   */
-  export function isNumber<T>(input: T): T extends number ? true : false;
-  export function isNumber(input: unknown): input is Number {
-    return typeof input === "number";
-  }
-
-  /**
-   * function인지 여부를 반환합니다.
-   *
-   * @param input 검증 대상 값
-   * @returns input이 function인지 여부
-   */
-  export function isFunction<T>(input: T): T extends Function ? true : false;
-  export function isFunction(input: unknown): input is Function {
-    return typeof input === "function";
-  }
-
-  /**
-   * 배열 혹은 객체의 값의 배열의 순서를 무작위로 섞어 반환합니다.
-   *
-   * @param collection 섞을 배열 혹은 객체
-   * @returns 배열인 경우 섞인 배열, 객체인 경우 객체의 값으로 이루어진 섞인 배열
-   */
-  export function shuffle<T>(collection: T[]): T[];
-  export function shuffle<T>(collection: Record<string, T>): T[];
-  export function shuffle<T>(collection: T[] | Record<string, T>) {
-    const result =
-      collection instanceof Array ? collection : Object.values(collection);
-
-    let currIndex = result.length;
-    let randomIndex;
-
-    while (currIndex !== 0) {
-      randomIndex = Math.floor(Math.random() * currIndex);
-      currIndex -= 1;
-      [result[currIndex], result[randomIndex]] = [
-        result[randomIndex],
-        result[currIndex],
-      ];
-    }
-
-    return result;
-  }
-
-  /**
-   * 객체에서 일부 속성 값만 추출해서 만든 새로운 객체를 반환합니다.
-   *
-   * @param object 대상 객체
-   * @param props 선택할 속성의 배열
-   * @returns 대상 객체 중 선택한 속성만으로 구성된 새로운 객체
-   */
-  export function pick<
-    T extends Record<string, unknown>,
-    U extends Partial<keyof T>
-  >(object: T, props: U[]): Pick<T, U> {
-    return props.reduce((acc, curr) => {
-      acc[curr] = object[curr];
-      return acc;
-    }, {} as Pick<T, U>);
-  }
-
-  /**
-   * 객체에서 일부 속성 값만 제거해서 만든 새로운 객체를 반환합니다.
-   *
-   * @param object 대상 객체
-   * @param props 제거할 속성의 배열
-   * @returns 대상 객체에서 선택한 속성을 제거한 새로운 객체
-   */
-  export function omit<
-    T extends Record<string, unknown>,
-    U extends Partial<keyof T>
-  >(object: T, props?: U[]): Omit<T, U> {
-    return props === undefined
-      ? object
-      : props.reduce((acc, curr) => {
-          delete acc[curr];
-          return acc;
-        }, object);
-  }
-
-  /**
-   * 함수의 연산에 메모이제이션을 적용해, 동일한 인자가 주어졌을 때 반복적인 연산을 생략하고 저장된 값을 반환하는 함수를 반환합니다.
-   * resolver 함수를 전달하지 않으면 자동으로 첫 번째 인자를 캐시 키로 활용합니다.
-   *
-   * @param func 메모이제이션을 적용할 함수
-   * @param resolver 캐시 키를 생성하는 함수
-   * @returns 메모이제이션이 적용된 함수
-   */
-  export function memoize<T extends unknown[], U>(
-    func: (...args: T) => U,
-    resolver?: (...args: T) => string
-  ): (...args: T) => U {
-    const cache = new Map();
-    const memoizedFunction = (...args: T) => {
-      const cacheKey = resolver !== undefined ? resolver(...args) : args[0];
-      if (cache.has(cacheKey)) return cache.get(cacheKey);
-
-      const result = func(...args);
-      cache.set(cacheKey, result);
-
-      return result;
-    };
-    return memoizedFunction;
-  }
-
-  /**
-   * 실행하기 전 정해진 시간을 대기하는 디바운스가 적용된 함수를 반환합니다.
-   * 지연 시간을 전달하지 않으면 지연되지 않고 바로 함수가 실행됩니다.
-   *
-   * @param func 디바운스를 적용할 함수
-   * @param wait 함수 실행 지연시간
-   * @returns 디바운스가 적용된 함수
-   */
-  export const debounce: <T extends unknown[], U>(
-    func: (...args: T) => U,
-    wait?: number
-  ) => (...args: T) => U = <T extends unknown[], U>(
-    func: (...args: T) => U,
-    wait = 0
-  ) => {
-    let timer: NodeJS.Timeout, result: U;
-
-    const debouncedFunction = (...args: T) => {
-      if (timer) clearTimeout(timer);
-      timer = setTimeout(() => {
-        result = func(...args);
-      }, wait);
-
-      return result;
-    };
-
-    return debouncedFunction;
-  };
-
-  /**
-   * 일정 시간 동안 함수를 한 번만 실행하는 쓰로틀이 적용된 함수를 반환합니다.
-   * 주기를 전달하지 않으면 함수가 호출 될 때마다 실행됩니다.
-   *
-   * @param func 쓰로틀을 적용할 함수
-   * @param wait 함수를 1회 실행할 주기
-   * @returns 쓰로틀이 적용된 함수
-   */
-  export const throttle: <T extends unknown[], U>(
-    func: (...args: T) => U,
-    wait?: number
-  ) => (...args: T) => U | undefined = <T extends unknown[], U>(
-    func: (...args: T) => U,
-    wait = 0
-  ) => {
-    let shouldRun = true;
-    let result: U;
-
-    const throttledFunction = (...args: T) => {
-      if (!shouldRun) return;
-
-      shouldRun = false;
-      setTimeout(() => {
-        result = func(...args);
-
-        shouldRun = true;
-      }, wait);
-
-      return result;
-    };
-
-    return throttledFunction;
-  };
-
-  /**
-   * 전달된 대상 요소가 아닌 다른 요소가 클릭되었을 때 콜백함수를 호출하는 이벤트 리스너를 추가합니다.
-   *
-   * @param targetElement 대상 요소
-   * @param handler 대상 요소가 아닌 곳이 클릭됐을 때 호출할 콜백 함수
-   */
-  export function clickOutside(
-    targetElement: Element,
-    handler: (event: MouseEvent) => void
-  ): void {
-    window.addEventListener("click", (e) => {
-      if (e.target === targetElement) return;
-      handler(e);
-    });
-  }
+  export const fetch = _fetch;
+  export const isNull = _isNull;
+  export const isNil = _isNil;
+  export const isNumber = _isNumber;
+  export const isFunction = _isFunction;
+  export const shuffle = _shuffle;
+  export const pick = _pick;
+  export const omit = _omit;
+  export const memoize = _memoize;
+  export const debounce = _debounce;
+  export const throttle = _throttle;
+  export const clickOutside = _clickOutside;
 }
 
 export default _;
